requirements.txt
# Requirements for OIC ADEI Streamlit Analytics App
# =================================================

# Core Streamlit and Web Framework
streamlit==1.28.1
streamlit-aggrid==0.3.4.post3

# Data Processing
pandas==2.1.3
numpy==1.24.3

# Database Connectivity
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Machine Learning & AI
scikit-learn==1.3.2
langchain==0.0.340
langchain-openai==0.0.2
openai==1.3.7

# Vector Store & Embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Text Processing
tiktoken==0.5.2

# Utility Libraries
python-dotenv==1.0.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-streamlit==1.0.0

# Optional: Advanced NLP
transformers==4.36.2

# Development
black==23.11.0
flake8==6.1.0